[NestedStack] Split the compute fleet stack into finer nested stacks for groups of queues
