+++ cli/src/pcluster/constants.py
@@ -95,7 +95,7 @@ EBS_VOLUME_TYPE_DEFAULT = "gp3"
 
 DEFAULT_MAX_COUNT = 10
 DEFAULT_MIN_COUNT = 0
-MAX_NUMBER_OF_QUEUES = 10
+MAX_NUMBER_OF_QUEUES = 101
 MAX_NUMBER_OF_COMPUTE_RESOURCES = 5
 
 MAX_EBS_COUNT = 5
